requirements.txt
flask
flask-cors
gunicorn
python-dotenv
requests
feedparser
google-generativeai
openai
schedule
sqlalchemy>=2.0.0
psycopg2-binary
alembic
pdfkit
weasyprint